Revision Update Validation & Soft Delete

+++ elearning/src/main/java/com/lawencon/elearning/service/ModulesServiceImpl.java
@@ -1,6 +1,7 @@
 package com.lawencon.elearning.service;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -34,17 +35,29 @@ public class ModulesServiceImpl extends BaseServiceImpl implements ModulesServic
 	public void deleteModuleById(String id, String idUser) throws Exception {
 		try {
 			begin();
-			modulesDao.deleteModuleById(id);
-			commit();
-		} catch(Exception e) {
-			e.printStackTrace();
-			if(e.getMessage().equals("ID Not Found")) {
-				throw new Exception("Id tidak ada");
+			if(validateDelete(idUser)) {
+				
+			} else {			
+				modulesDao.deleteModuleById(id);
 			}
-			begin();
-			updateIsActive(id, idUser);
-			commit();
+			commit();			
+		} catch(Exception e) {
+			e.getMessage();
+			rollback();
 		}
+//		try {
+//			begin();
+//			modulesDao.deleteModuleById(id);
+//			commit();
+//		} catch(Exception e) {
+//			e.printStackTrace();
+//			if(e.getMessage().equals("ID Not Found")) {
+//				throw new Exception("Id tidak ada");
+//			}
+//			begin();
+//			updateIsActive(id, idUser);
+//			commit();
+//		}
 	}
 
 	@Override
@@ -60,24 +73,48 @@ public class ModulesServiceImpl extends BaseServiceImpl implements ModulesServic
 	private void validateInsert(Modules module) throws Exception {
 		if(module.getCode() == null || module.getCode().trim().equals("")) {
 			throw new Exception("Kode Modul tidak boleh kosong");
-		} else if(module.getModuleName() == null || module.getModuleName().trim().equals("")) {
-			throw new Exception("Nama Modul tidak boleh kosong");
-		}
+		} else {
+			Modules mod = getModuleByCode(module.getCode());
+			if(mod != null) {
+				throw new Exception("Kode Modul sudah ada");
+			}
+			if(module.getModuleName() == null || module.getModuleName().trim().equals("")) {
+				throw new Exception("Nama Modul tidak boleh kosong");
+			}
+		} 
 	}
 
 	private void validateUpdate(Modules module) throws Exception {
 		if(module.getId() == null || module.getId().trim().equals("")) {
 			throw new Exception("Id Modul tidak boleh kosong");
-		} else if(module.getCode() == null || module.getCode().trim().equals("")) {
+		}
+		if(module.getCode() == null || module.getCode().trim().equals("")) {
 			throw new Exception("Kode Modul tidak boleh kosong");
-		} else if(module.getModuleName() == null || module.getModuleName().trim().equals("")) {
-			throw new Exception("Nama Modul tidak boleh kosong");
+		} else {
+			Modules mod = getModuleById(module.getId());
+			if(mod != null) {
+				if(!mod.getCode().equals(module.getCode())) {
+					throw new Exception("Kode Modul sudah ada");
+				}
+			}
+			if(module.getModuleName() == null || module.getModuleName().trim().equals("")) {
+				throw new Exception("Nama Modul tidak boleh kosong");
+			}			
 		}
 	}
 
 	@Override
 	public void updateIsActive(String id, String idUser) throws Exception {
 		modulesDao.updateIsActive(id, idUser);
+	}
+	
+	private boolean validateDelete(String id) throws Exception {
+		List<?> listObj = modulesDao.validateDeleteModule(id);
+		listObj.forEach(System.out::println);
+		List<?> list =  listObj.stream().filter(val -> val != null)
+				.collect(Collectors.toList());
+		System.out.println(list.size());
+		return list.size() > 0 ? true : false;
 	}
 
 }
